Consumers/requirements.txt
kafka-python
confluent-kafka
aiokafka